refactoring sliderfield demo
