+++ docs/src/pages/components/sliderfield/react.mdx
@@ -43,7 +43,7 @@ import '@aws-amplify/ui-react/styles.css';
   />
 </Example>
 
-The `Slider` will render an input for the thumb when used within a form to ensure that value will be submiited with form data.
+The `Slider` will render an input for the thumb when used within a form to ensure that value will be submitted with form data.
 
 ```jsx
 import { Button, SliderField } from '@aws-amplify/ui-react';
